Update codebase to current master and establish p3 infrastructure.
This commit puts in place the infrastructure for eventually porting
P3 into E3SM master.  It updates the namelist options to include
P3 specific variables and to allow for a P3 microphysics option.

It updates schemes that are dependent on the microphysics scheme
option, such as physpkg and radiation to also have the option for
the p3 microphysics option.

There are a few TODO items that must be addressed once the actual
P3 code has been added to the model, they can be found by grepping
for "AaronDonahue TODO" in the cam codebase.

This commit is also built on the most recent version of master,
updating the E3SM code base in the SCREAM fork.