Hommexx: Remove redundant HOMMEXX_ENABLE_GPU definition.
This symbol was being defined in both Hommexx_config.h and config.h. The
distinction between these two is subtle but definite. The first is for symbols
used only in the Hommexx code. The second is for symbols used in both Hommexx
and Homme, the latter including Homme dycores built for BFB testing with
Hommexx.

Since the symbol is redundantly defined (leading to a preprocessor warning about
this), one must be removed. It's used in Homme, so the Hommexx_config.h.in
instance should be removed.